+++ egs/musdb18/common/src/dataset.py
@@ -1,4 +1,5 @@
 import os
+import random
 
 import torch
 import torchaudio
@@ -96,10 +97,12 @@ class WaveDataset(MUSDB18Dataset):
         return len(self.json_data)
 
 class WaveTrainDataset(WaveDataset):
-    def __init__(self, musdb18_root, sr=SAMPLE_RATE_MUSDB18, samples=4*SAMPLE_RATE_MUSDB18, overlap=None, sources=__sources__, target=None, threshold=THRESHOLD_POWER):
+    def __init__(self, musdb18_root, sr=SAMPLE_RATE_MUSDB18, samples=4*SAMPLE_RATE_MUSDB18, overlap=None, sources=__sources__, target=None, include_valid=False):
+        """
+        Args:
+            include_valid <bool>: Include validation data for training.
+        """
         super().__init__(musdb18_root, sr=sr, sources=sources, target=target)
-        
-        assert_sample_rate(sr)
 
         valid_txt_path = os.path.join(musdb18_root, 'validation.txt')
         train_txt_path = os.path.join(musdb18_root, 'train.txt')
@@ -113,11 +116,10 @@ class WaveTrainDataset(WaveDataset):
             for line in f:
                 name = line.strip()
 
-                if name in valid_lst:
+                if (not include_valid) and name in valid_lst:
                     continue
+                
                 names.append(name)
-
-        self.threshold = threshold
 
         if overlap is None:
             overlap = samples // 2
@@ -167,8 +169,6 @@ class WaveTrainDataset(WaveDataset):
 class WaveEvalDataset(WaveDataset):
     def __init__(self, musdb18_root, sr=SAMPLE_RATE_MUSDB18, max_samples=4*SAMPLE_RATE_MUSDB18, sources=__sources__, target=None):
         super().__init__(musdb18_root, sr=sr, sources=sources, target=target)
-        
-        assert_sample_rate(sr)
 
         valid_txt_path = os.path.join(musdb18_root, 'validation.txt')
 
@@ -231,8 +231,6 @@ class WaveEvalDataset(WaveDataset):
 class WaveTestDataset(WaveDataset):
     def __init__(self, musdb18_root, sr=SAMPLE_RATE_MUSDB18, sources=__sources__, target=None):
         super().__init__(musdb18_root, sr=sr, sources=sources, target=target)
-        
-        assert_sample_rate(sr)
 
         test_txt_path = os.path.join(musdb18_root, 'test.txt')
 
@@ -275,8 +273,6 @@ class WaveTestDataset(WaveDataset):
 class SpectrogramDataset(WaveDataset):
     def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, sources=__sources__, target=None):
         super().__init__(musdb18_root, sr=sr, sources=sources, target=target)
-        
-        assert_sample_rate(sr)
 
         if hop_size is None:
             hop_size = fft_size // 2
@@ -335,7 +331,7 @@ class SpectrogramDataset(WaveDataset):
         return mixture, target, T, name
 
 class SpectrogramTrainDataset(SpectrogramDataset):
-    def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, samples=4*SAMPLE_RATE_MUSDB18, overlap=None, sources=__sources__, target=None, threshold=THRESHOLD_POWER):
+    def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, samples=4*SAMPLE_RATE_MUSDB18, overlap=None, sources=__sources__, target=None, include_valid=False):
         super().__init__(musdb18_root, fft_size=fft_size, hop_size=hop_size, window_fn=window_fn, normalize=normalize, sr=sr, sources=sources, target=target)
         
         assert_sample_rate(sr)
@@ -352,11 +348,10 @@ class SpectrogramTrainDataset(SpectrogramDataset):
             for line in f:
                 name = line.strip()
 
-                if name in valid_lst:
+                if (not include_valid) and name in valid_lst:
                     continue
+                
                 names.append(name)
-
-        self.threshold = threshold
 
         if overlap is None:
             overlap = samples // 2
@@ -406,8 +401,6 @@ class SpectrogramTrainDataset(SpectrogramDataset):
 class SpectrogramEvalDataset(SpectrogramDataset):
     def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, max_samples=10*SAMPLE_RATE_MUSDB18, sources=__sources__, target=None):
         super().__init__(musdb18_root, fft_size=fft_size, hop_size=hop_size, window_fn=window_fn, normalize=normalize, sr=sr, sources=sources, target=target)
-
-        assert_sample_rate(sr)
 
         valid_txt_path = os.path.join(musdb18_root, 'validation.txt')
 
@@ -473,8 +466,6 @@ class SpectrogramTestDataset(SpectrogramDataset):
     def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, max_samples=10*SAMPLE_RATE_MUSDB18, sources=__sources__, target=None):
         super().__init__(musdb18_root, fft_size=fft_size, hop_size=hop_size, window_fn=window_fn, normalize=normalize, sr=sr, sources=sources, target=target)
 
-        assert_sample_rate(sr)
-
         test_txt_path = os.path.join(musdb18_root, 'test.txt')
 
         names = []
@@ -534,6 +525,378 @@ class SpectrogramTestDataset(SpectrogramDataset):
         mixture, sources, T, name = super().__getitem__(idx)
         
         return mixture, sources, T, name
+
+"""
+    Augmentation dataset
+"""
+class AugmentationWaveTrainDataset(WaveDataset):
+    def __init__(self, musdb18_root, sr=SAMPLE_RATE_MUSDB18, duration=4, overlap=None, samples_per_epoch=None, sources=__sources__, target=None, include_valid=False, augmentation=None):
+        """
+        Args:
+            musdb18_root <int>: Path to MUSDB or MUSDB-HQ
+            sr: Sampling frequency. Default: 44100
+            sources <list<str>>: Sources included in mixture
+            target <str> or <list<str>>: Target source(s)
+        """
+        super().__init__(
+            musdb18_root,
+            sr=SAMPLE_RATE_MUSDB18, # WaveDataset's sr is expected SAMPLE_RATE_MUSDB18
+            sources=sources,
+            target=target
+        )
+
+        valid_txt_path = os.path.join(musdb18_root, 'validation.txt')
+        train_txt_path = os.path.join(musdb18_root, 'train.txt')
+
+        with open(valid_txt_path, 'r') as f:
+            valid_lst = [line.strip() for line in f]
+
+        names = []
+
+        with open(train_txt_path, 'r') as f:
+            for line in f:
+                name = line.strip()
+
+                if (not include_valid) and name in valid_lst:
+                    continue
+                
+                names.append(name)
+
+        self.sr = sr
+        self.samples = int(duration * sr)
+        self.augmentation = augmentation
+
+        self.tracks = []
+
+        if augmentation:
+            total_duration = 0
+
+            for trackID, name in enumerate(names):
+                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
+                audio_info = torchaudio.info(mixture_path)
+                track_sr = audio_info.sample_rate
+                track_samples = audio_info.num_frames
+
+                track = {
+                    'name': name,
+                    'samples_original': track_samples,
+                    'path': {
+                        'mixture': mixture_path
+                    }
+                }
+                
+                for source in sources:
+                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
+                
+                self.tracks.append(track)
+
+                track_duration = track_samples / track_sr
+                total_duration += track_duration
+
+            if samples_per_epoch is None:
+                samples_per_epoch = int(total_duration / duration)
+
+            self.samples_per_epoch = samples_per_epoch
+            self.json_data = None
+        else:
+            samples_original = int(self.samples * SAMPLE_RATE_MUSDB18 / sr)
+
+            if overlap is None:
+                overlap = samples_original // 2
+            self.samples_per_epoch = None
+
+            for trackID, name in enumerate(names):
+                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
+                audio_info = torchaudio.info(mixture_path)
+                track_sr = audio_info.sample_rate
+                track_samples = audio_info.num_frames
+
+                track = {
+                    'name': name,
+                    'samples_original': track_samples,
+                    'path': {
+                        'mixture': mixture_path
+                    }
+                }
+
+                for source in sources:
+                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
+                self.tracks.append(track)
+
+                for start in range(0, track_samples, self.samples - overlap):
+                    if start + self.samples >= track_samples:
+                        break
+                    data = {
+                        'trackID': trackID,
+                        'start': start,
+                        'samples_original': samples_original,
+                    }
+                    self.json_data.append(data)
+
+        if sr != SAMPLE_RATE_MUSDB18:
+            self.pre_resampler = torchaudio.transforms.Resample(SAMPLE_RATE_MUSDB18, sr)
+        else:
+            self.pre_resampler = None
+
+    def __getitem__(self, idx):
+        """
+        Returns:
+            mixture <torch.Tensor>: Tensor with shape (1, n_mics, T)  if `target` is list, otherwise (n_mics, T) 
+            target <torch.Tensor>: Tensor with shape (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
+        """
+        if self.augmentation:
+            mixture, target = self._getitem_augmentation()
+        else:
+            raise NotImplementedError("Implement _getitem()")
+        
+        if self.pre_resampler is not None:
+            mixture_channels, target_channels = mixture.size()[:-1], target.size()[:-1]
+            mixture, target = mixture.reshape(-1, mixture.size(-1)), target.reshape(-1, target.size(-1))
+
+            mixture = self.pre_resampler(mixture)
+            target = self.pre_resampler(target)
+
+            mixture, target = mixture.reshape(*mixture_channels, mixture.size(-1)), target.reshape(*target_channels, target.size(-1))
+
+        return mixture, target
+
+    def _getitem_augmentation(self):
+        """
+        Returns time domain signals
+        Returns:
+            mixture <torch.Tensor>: (1, n_mics, T) if `target` is list, otherwise (n_mics, T)
+            target <torch.Tensor>: (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
+            name <str>: Artist and title of track
+        """
+        n_tracks = len(self.tracks)
+        track_indices = random.choices(range(n_tracks), k=len(self.sources))
+
+        sources = []
+
+        for _source, trackID in zip(self.sources, track_indices):
+            track = self.tracks[trackID]
+            source_path = track['path'][_source]
+            track_samples = track['samples_original']
+
+            start = random.randint(0, track_samples - self.samples - 1)
+            source, _ = torchaudio.load(source_path, frame_offset=start, num_frames=self.samples)
+
+            # Apply augmentation
+            source = self.augmentation(source)
+            sources.append(source.unsqueeze(dim=0))
+        
+        if type(self.target) is list:
+            target = []
+            for _target in self.target:
+                source_idx = self.sources.index(_target)
+                _target = sources[source_idx]
+                target.append(_target)
+            target = torch.cat(target, dim=0)
+
+            sources = torch.cat(sources, dim=0)
+            mixture = sources.sum(dim=0, keepdim=True)
+        else:
+            source_idx = self.sources.index(self.target)
+            target = sources[source_idx]
+            target = target.squeeze(dim=0)
+
+            sources = torch.cat(sources, dim=0)
+            mixture = sources.sum(dim=0)
+
+        return mixture, target
+
+    def __len__(self):
+        return self.samples_per_epoch
+
+class AugmentationSpectrogramTrainDataset(SpectrogramDataset):
+    """
+    Training dataset that returns randomly selected mixture spectrograms.
+    """
+    def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, patch_samples=6*SAMPLE_RATE_MUSDB18, overlap=None, samples_per_epoch=None, sources=__sources__, target=None, include_valid=False, augmentation=None):
+        super().__init__(musdb18_root, fft_size=fft_size, hop_size=hop_size, window_fn=window_fn, normalize=normalize, sr=sr, sources=sources, target=target)
+        
+        valid_txt_path = os.path.join(musdb18_root, 'validation.txt')
+        train_txt_path = os.path.join(musdb18_root, 'train.txt')
+
+        with open(valid_txt_path, 'r') as f:
+            valid_lst = [line.strip() for line in f]
+
+        names = []
+
+        with open(train_txt_path, 'r') as f:
+            for line in f:
+                name = line.strip()
+
+                if (not include_valid) and name in valid_lst:
+                    continue
+                
+                names.append(name)
+        
+        self.patch_samples = patch_samples
+
+        self.augmentation = augmentation
+
+        self.tracks = []
+
+        if augmentation:
+            duration = patch_samples / sr
+            total_duration = 0
+
+            for trackID, name in enumerate(names):
+                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
+                audio_info = torchaudio.info(mixture_path)
+                sr = audio_info.sample_rate
+                track_samples = audio_info.num_frames
+
+                track = {
+                    'name': name,
+                    'samples': track_samples,
+                    'path': {
+                        'mixture': mixture_path
+                    }
+                }
+                
+                for source in sources:
+                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
+                
+                self.tracks.append(track)
+
+                track_duration = track_samples / sr
+                total_duration += track_duration
+
+            if samples_per_epoch is None:
+                samples_per_epoch = int(total_duration / duration)
+
+            self.samples_per_epoch = samples_per_epoch
+            self.json_data = None
+        else:
+            if overlap is None:
+                overlap = patch_samples // 2
+            
+            self.samples_per_epoch = None
+
+            for trackID, name in enumerate(names):
+                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
+                audio_info = torchaudio.info(mixture_path)
+                sr = audio_info.sample_rate
+                track_samples = audio_info.num_frames
+
+                track = {
+                    'name': name,
+                    'samples': track_samples,
+                    'path': {
+                        'mixture': mixture_path
+                    }
+                }
+
+                for source in sources:
+                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
+                
+                self.tracks.append(track)
+
+                for start in range(0, track_samples, patch_samples - overlap):
+                    if start + patch_samples >= track_samples:
+                        break
+                    data = {
+                        'trackID': trackID,
+                        'start': start,
+                        'samples': patch_samples,
+                    }
+                    self.json_data.append(data)
+
+    def __getitem__(self, idx):
+        """
+        Returns:
+            mixture <torch.Tensor>: Complex tensor with shape (1, n_mics, n_bins, n_frames)  if `target` is list, otherwise (n_mics, n_bins, n_frames) 
+            target <torch.Tensor>: Complex tensor with shape (len(target), n_mics, n_bins, n_frames) if `target` is list, otherwise (n_mics, n_bins, n_frames)
+        """
+        if self.augmentation:
+            mixture, target = self._getitem_augmentation()
+        else:
+            mixture, target = self._getitem(idx)
+        
+        n_dims = mixture.dim()
+
+        if n_dims > 2:
+            mixture_channels = mixture.size()[:-1]
+            target_channels = target.size()[:-1]
+            mixture = mixture.reshape(-1, mixture.size(-1))
+            target = target.reshape(-1, target.size(-1))
+
+        mixture = torch.stft(mixture, n_fft=self.fft_size, hop_length=self.hop_size, window=self.window, normalized=self.normalize, return_complex=True) # (1, n_mics, n_bins, n_frames) or (n_mics, n_bins, n_frames)
+        target = torch.stft(target, n_fft=self.fft_size, hop_length=self.hop_size, window=self.window, normalized=self.normalize, return_complex=True) # (len(sources), n_mics, n_bins, n_frames) or (n_mics, n_bins, n_frames)
+        
+        if n_dims > 2:
+            mixture = mixture.reshape(*mixture_channels, *mixture.size()[-2:])
+            target = target.reshape(*target_channels, *target.size()[-2:])
+
+        return mixture, target
+    
+    def __len__(self):
+        if self.augmentation:
+            return self.samples_per_epoch
+        else:
+            source = self.sources[0]
+            
+            return len(self.json_data[source])
+    
+    def _getitem(self, idx):
+        """
+        Returns time domain signals
+        Args:
+            idx <int>: index
+        Returns:
+            mixture <torch.Tensor>: (1, n_mics, T) if `target` is list, otherwise (n_mics, T)
+            target <torch.Tensor>: (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
+        """
+        mixture, target, _, _ = super().__getitem__(idx)
+
+        return mixture, target
+    
+    def _getitem_augmentation(self):
+        """
+        Returns time domain signals
+        Returns:
+            mixture <torch.Tensor>: (1, n_mics, T) if `target` is list, otherwise (n_mics, T)
+            target <torch.Tensor>: (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
+            name <str>: Artist and title of track
+        """
+        n_tracks = len(self.tracks)
+        track_indices = random.choices(range(n_tracks), k=len(self.sources))
+
+        sources = []
+
+        for _source, trackID in zip(self.sources, track_indices):
+            track = self.tracks[trackID]
+            source_path = track['path'][_source]
+            track_samples = track['samples']
+
+            start = random.randint(0, track_samples - self.patch_samples - 1)
+            source, _ = torchaudio.load(source_path, frame_offset=start, num_frames=self.patch_samples)
+
+            # Apply augmentation
+            source = self.augmentation(source)
+            sources.append(source.unsqueeze(dim=0))
+        
+        if type(self.target) is list:
+            target = []
+            for _target in self.target:
+                source_idx = self.sources.index(_target)
+                _target = sources[source_idx]
+                target.append(_target)
+            target = torch.cat(target, dim=0)
+
+            sources = torch.cat(sources, dim=0)
+            mixture = sources.sum(dim=0, keepdim=True)
+        else:
+            source_idx = self.sources.index(self.target)
+            target = sources[source_idx]
+            target = target.squeeze(dim=0)
+
+            sources = torch.cat(sources, dim=0)
+            mixture = sources.sum(dim=0)
+
+        return mixture, target
 
 """
     Data loader
