+++ egs/musdb18/d3net/src/adhoc_dataset.py
@@ -1,5 +1,4 @@
 import os
-import random
 
 import torch
 import torchaudio
@@ -12,183 +11,6 @@ __sources__ = ['bass', 'drums', 'other', 'vocals']
 SAMPLE_RATE_MUSDB18 = 44100
 EPS = 1e-12
 THRESHOLD_POWER = 1e-5
-
-class SpectrogramTrainDataset(SpectrogramDataset):
-    """
-    Training dataset that returns randomly selected mixture spectrograms.
-    In accordane with "D3Net: Densely connected multidilated DenseNet for music source separation," training dataset includes all 100 tracks.
-    """
-    def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, patch_samples=4*SAMPLE_RATE_MUSDB18, overlap=None, samples_per_epoch=None, sources=__sources__, target=None, augmentation=None):
-        super().__init__(musdb18_root, fft_size=fft_size, hop_size=hop_size, window_fn=window_fn, normalize=normalize, sr=sr, sources=sources, target=target)
-        
-        train_txt_path = os.path.join(musdb18_root, 'train.txt')
-
-        with open(train_txt_path, 'r') as f:
-            names = [line.strip() for line in f]
-        
-        self.patch_samples = patch_samples
-
-        self.augmentation = augmentation
-
-        self.tracks = []
-
-        if augmentation:
-            duration = patch_samples / sr
-            total_duration = 0
-
-            for trackID, name in enumerate(names):
-                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
-                audio_info = torchaudio.info(mixture_path)
-                sr = audio_info.sample_rate
-                track_samples = audio_info.num_frames
-
-                track = {
-                    'name': name,
-                    'samples': track_samples,
-                    'path': {
-                        'mixture': mixture_path
-                    }
-                }
-                
-                for source in sources:
-                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
-                
-                self.tracks.append(track)
-
-                track_duration = track_samples / sr
-                total_duration += track_duration
-
-            if samples_per_epoch is None:
-                samples_per_epoch = int(total_duration / duration)
-
-            self.samples_per_epoch = samples_per_epoch
-            self.json_data = None
-        else:
-            if overlap is None:
-                overlap = patch_samples // 2
-            self.samples_per_epoch = None
-
-            for trackID, name in enumerate(names):
-                mixture_path = os.path.join(musdb18_root, 'train', name, "mixture.wav")
-                audio_info = torchaudio.info(mixture_path)
-                sr = audio_info.sample_rate
-                track_samples = audio_info.num_frames
-
-                track = {
-                    'name': name,
-                    'samples': track_samples,
-                    'path': {
-                        'mixture': mixture_path
-                    }
-                }
-
-                for source in sources:
-                    track['path'][source] = os.path.join(musdb18_root, 'train', name, "{}.wav".format(source))
-                self.tracks.append(track)
-
-                for start in range(0, track_samples, patch_samples - overlap):
-                    if start + patch_samples >= track_samples:
-                        break
-                    data = {
-                        'trackID': trackID,
-                        'start': start,
-                        'samples': patch_samples,
-                    }
-                    self.json_data.append(data)
-    
-    def __getitem__(self, idx):
-        """
-        Returns:
-            mixture <torch.Tensor>: Complex tensor with shape (1, n_mics, n_bins, n_frames)  if `target` is list, otherwise (n_mics, n_bins, n_frames) 
-            target <torch.Tensor>: Complex tensor with shape (len(target), n_mics, n_bins, n_frames) if `target` is list, otherwise (n_mics, n_bins, n_frames)
-        """
-        if self.augmentation:
-            mixture, target = self._getitem_augmentation()
-        else:
-            mixture, target = self._getitem(idx)
-        
-        n_dims = mixture.dim()
-
-        if n_dims > 2:
-            mixture_channels = mixture.size()[:-1]
-            target_channels = target.size()[:-1]
-            mixture = mixture.reshape(-1, mixture.size(-1))
-            target = target.reshape(-1, target.size(-1))
-
-        mixture = torch.stft(mixture, n_fft=self.fft_size, hop_length=self.hop_size, window=self.window, normalized=self.normalize, return_complex=True) # (1, n_mics, n_bins, n_frames) or (n_mics, n_bins, n_frames)
-        target = torch.stft(target, n_fft=self.fft_size, hop_length=self.hop_size, window=self.window, normalized=self.normalize, return_complex=True) # (len(sources), n_mics, n_bins, n_frames) or (n_mics, n_bins, n_frames)
-        
-        if n_dims > 2:
-            mixture = mixture.reshape(*mixture_channels, *mixture.size()[-2:])
-            target = target.reshape(*target_channels, *target.size()[-2:])
-
-        return mixture, target
-    
-    def __len__(self):
-        if self.augmentation:
-            return self.samples_per_epoch
-        else:
-            source = self.sources[0]
-            
-            return len(self.json_data[source])
-
-    def _getitem(self, idx):
-        """
-        Returns time domain signals
-        Args:
-            idx <int>: index
-        Returns:
-            mixture <torch.Tensor>: (1, n_mics, T) if `target` is list, otherwise (n_mics, T)
-            target <torch.Tensor>: (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
-        """
-        mixture, target, _, _ = super().__getitem__(idx)
-
-        return mixture, target
-    
-    def _getitem_augmentation(self):
-        """
-        Returns time domain signals
-        Returns:
-            mixture <torch.Tensor>: (1, n_mics, T) if `target` is list, otherwise (n_mics, T)
-            target <torch.Tensor>: (len(target), n_mics, T) if `target` is list, otherwise (n_mics, T)
-            name <str>: Artist and title of track
-        """
-        n_tracks = len(self.tracks)
-        track_indices = random.choices(range(n_tracks), k=len(self.sources))
-
-        sources = []
-
-        for _source, trackID in zip(self.sources, track_indices):
-            track = self.tracks[trackID]
-            source_path = track['path'][_source]
-            track_samples = track['samples']
-
-            start = random.randint(0, track_samples - self.patch_samples - 1)
-            source, _ = torchaudio.load(source_path, frame_offset=start, num_frames=self.patch_samples)
-
-            # Apply augmentation
-            source = self.augmentation(source)
-            sources.append(source.unsqueeze(dim=0))
-        
-        if type(self.target) is list:
-            target = []
-            for _target in self.target:
-                source_idx = self.sources.index(_target)
-                _target = sources[source_idx]
-                target.append(_target)
-            target = torch.cat(target, dim=0)
-
-            sources = torch.cat(sources, dim=0)
-            mixture = sources.sum(dim=0, keepdim=True)
-        else:
-            source_idx = self.sources.index(self.target)
-            target = sources[source_idx]
-            target = target.squeeze(dim=0)
-
-            sources = torch.cat(sources, dim=0)
-            mixture = sources.sum(dim=0)
-
-        return mixture, target
 
 class SpectrogramEvalDataset(SpectrogramDataset):
     def __init__(self, musdb18_root, fft_size, hop_size=None, window_fn='hann', normalize=False, sr=SAMPLE_RATE_MUSDB18, patch_size=256, max_samples=None, sources=__sources__, target=None):
